+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the "aiXcoder" extension will be documented in this file.
 
 ### Fixed
 - Fix a bug that prevents local server launching automatically.
+- Fix a bug when there is quote in comment.
 
 ## [0.2.2] - 2019-11-18